Make CurrencyId accessible through primitives
